Move espresso streamer out of arbstate
